Multi-select: move handlers to block ref callback
